+++ src/main/coerce.c
@@ -127,7 +127,7 @@ int attribute_hidden IntegerFromReal(double x, int *warn)
         *warn |= WARN_NA;
         return NA_INTEGER;
     }
-    return x;
+    return (int)x;
 }
 
 int attribute_hidden IntegerFromComplex(Rcomplex x, int *warn)
@@ -142,7 +142,7 @@ int attribute_hidden IntegerFromComplex(Rcomplex x, int *warn)
     }
     if (x.i != 0)
         *warn |= WARN_IMAG;
-    return x.r;
+    return (int)x.r;
 }
 
 int attribute_hidden IntegerFromString(SEXP x, int *warn)
